Bigger map, Powersystem
The map is now a little bigger and it is more clear of where you need to go.

There is a new item the powersystem. You will need this item to use the power system in the future. With this item you can collect new powers and use them. (this is the idea at the moment may change. it is also not fully implemented)