Dialogue, Power system, new level
Fixed a issue with not displaying the correct dialogue text.
Added correct text to the first dialogue cutscene (when you pickup the powermachine)
You now have to pickup the powersystem item before you can use the power
Added a new level to learn how to use the timecontrol power. with a reset function