Cutscene, fixed bugs and added new level
A new cutscene is added that tells you about the story of why everything is like how to world is.

Fixed some bugs that where apearing.

Added a new level at the end of the game where you need to do a little of pakour to go to the end.