Laser system, respawn things
I fixed the laser system so you can now shoot a laser with power 3 and damage the enemy with it. by shooting the enemy 5 times he will die and a way out of the arena will spawn so you can escape.
i also fixed the camera and wall so you will now see thru the normal camera when you respawn and can walk balk in the arena normal to go further in the fight.